EliminarConMensaje

+++ src/Pages/VistaClienteAdmin.js
@@ -28,9 +28,14 @@ const VistaClientesAdmin = () => {
     },[])
 
     const deleteRegistrarClientes = async (id) =>{
-        await axios.delete(`http://127.0.0.1:8000/api/Usuario/${id}`)
-        todosClientes()
-    }
+        var opcion = window.confirm("¿Estás Seguro que deseas Eliminar a este Usuario?");
+        if (opcion === true) {
+            await axios.delete(`http://127.0.0.1:8000/api/Usuario/${id}`)
+            todosClientes()
+            window.location.reload();
+        }
+      };
+
     
 
   return (
